+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed data integrity context to `https://w3id.org/security/data-integrity/v2`.
 - Use `eddsa-jcs-2022` cryptosuite when generating minisign identity proofs.
 
+### Fixed
+
+- Use correct verification method ID when creating identity proofs.
+
 ## [4.11.0] - 2025-10-09
 
 ### Added